Replacing files from common folde into extension
