+++ src/common/extensionMessaging.ts
@@ -13,7 +13,6 @@ export enum ExtensionMessage {
     GET_PACKAGER_PORT,
     SEND_TELEMETRY,
     OPEN_FILE_AT_LOCATION,
-    START_EXPONENT_PACKAGER,
     SHOW_INFORMATION_MESSAGE,
     LAUNCH,
 }
