+++ src/extension/android/packageNameResolver.ts
