+++ src/extension/exponent/exponentHelper.d.ts
