+++ src/extension/ios/iOSDebugModeManager.ts
