+++ test/extension/ios/iOSPlatform.test.ts
@@ -3,7 +3,7 @@
 
 import * as assert from "assert";
 
-import { IOSPlatform } from "../../../src/common/ios/iOSPlatform";
+import { IOSPlatform } from "../../../src/extension/ios/iOSPlatform";
 // import { IRunOptions } from "../../../src/common/launchArgs";
 
 import "should";
